The Great JavaScript Purge of 2023.
- Removes partials and dated scripts. Keeps script for navigation and
  moves it to main.js.
- Removes Gulp because phooey.
- Removes unneeded dependencies from NPM package list.
- Removes python scripts, even though we like python just fine.
