+++ _layouts/default.html
@@ -12,9 +12,6 @@
 
 		{% include footer.html %}
 
-		<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
-		<script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.19.1/moment.min.js" type="text/javascript" charset="utf-8"></script>
-		<script src="{{ '/assets/js/lib/hammer.min.js' | relative_url }}" type="text/javascript" charset="utf-8"></script>
 		<script src="{{ '/assets/js/main.js' | relative_url }}" type="text/javascript" charset="utf-8"></script>
         {% if page.title == 'Search' %}
 		<script src="https://unpkg.com/lunr/lunr.js" type="text/javascript" charset="utf-8"></script>
