--- assets/python/wrap_strings_in_html.py
@@ -1,13 +1,0 @@
-# wrap_strings_in_html.py
-
-title = "The Scholars' Lab"
-elem = 'span'
-
-def wrapLettersInHTML(string, elem):
-	title_letters = list(string.replace(' ', ''))
-	elems = ["<{e}>{t}</{e}>".format(e=elem, t=t) for t in title_letters]
-	return elems
-
-spans = wrapLettersInHTML(title, elem)
-for span in spans:
-	print(span)
